Added babel.py module for babel stuff.
